Provide default `Resource[]` <-> `ResourceFormState` conversion
